Fix the inconsistent info of availability zone
If the default availability zone is replaced with the
newly-created one, then it causes the info of availability
zone to be inconsistent for every retrieval in the process
of launching new VM instance.

This bug fix will not work if the availability zone is not
specified (e.g. via CLI of nova boot).

Closes-Bug: #1390033
Change-Id: I794b262e8f1e4adf10e3263e36a44c3a02008cb6